+++ src/theme/select.ts
@@ -10,19 +10,6 @@
 import { defuFn } from 'defu'
 import input from './input'
 import { buttonGroupVariantWithRoot } from './button-group'
-// import colorDropDownItem from './tools/color-drop-down-item'
-
-// @todo fix this
-const defSize = {
-  label: 'h-9 ps-2 pe-3 text-sm gap-2',
-  item: 'h-9 ps-3 pe-3 text-sm gap-2',
-  itemLeadingIcon: 'size-5',
-  itemLeadingAvatarSize: '2xs',
-  itemLeadingChip: 'size-3 not-group-data-reka-collection-item:ps-2.5',
-  itemLeadingChipSize: 'sm',
-  itemTrailingIcon: 'size-3',
-  trailingIcon: 'size-lg2'
-}
 
 export default () => {
   return defuFn(
@@ -55,98 +42,156 @@ export default () => {
         ].join(' '),
         value: 'truncate pointer-events-none',
         placeholder: 'truncate text-(--ui-color-design-plain-na-content-secondary)',
-        // @todo fix this
-        arrow: 'fill-base-master/10',
-        // @todo fix this
         content: [
-          'w-(--reka-select-trigger-width)',
-          // 'max-h-60',
-          'min-w-fit',
-          // 'h-(--reka-popper-available-height)',
-          'bg-white',
-          'shadow-md rounded-2xs ring ring-base-300',
-          'overflow-hidden',
-          'data-[state=open]:animate-[scale-in_100ms_ease-out] data-[state=closed]:animate-[scale-out_100ms_ease-in]',
-          'origin-(--reka-select-content-transform-origin)',
-          'flex flex-col',
+          'context-light',
+          '',
+          'bg-(--popup-window-background-color)',
+          'shadow-(--popup-window-box-shadow)',
+          'rounded-(--popup-window-border-radius) will-change-[opacity]',
+          'motion-safe:data-[state=open]:animate-[scale-in_100ms_ease-out] motion-safe:data-[state=closed]:animate-[scale-out_100ms_ease-in]',
+          'origin-(--reka-dropdown-menu-content-transform-origin)',
+          'font-[family-name:var(--ui-font-family-primary)]',
+          'relative',
+          'isolate',
+          'px-0 py-(--menu-popup-padding)',
           'pointer-events-auto'
         ].join(' '),
-        // @todo fix this
         viewport: [
           'relative',
-          'divide-y divide-base-master/10',
-          'scroll-py-1',
-          'overflow-y-auto',
-          'flex-1'
+          'w-[240px] max-h-[40vh]',
+          'overflow-x-hidden overflow-y-auto scrollbar-thin' // scrollbar-transparent
         ].join(' '),
-        group: 'p-1 isolate',
-        // @todo fix this
-        empty: 'py-2 text-center text-sm text-base-500',
-        // @todo fix this
+        arrow: 'fill-(--popup-window-background-color)', // for content bottom|top::start -> ml-[12px]
+        group: 'grid',
+        // @todo remove this
+        // empty: 'py-2 text-center text-sm text-base-500',
         label: [
-          'flex items-center',
-          'font-semibold text-base-900'
+          'w-full min-w-[195px] h-(--popup-window-delimiter-section-height)',
+          'px-[18px] mt-(--menu-item-block-stack-space)',
+          'flex flex-row rtl:flex-row-reverse items-center',
+          'select-none outline-none whitespace-nowrap',
+          'text-start',
+          'text-(length:--popup-window-delimiter-font-size)',
+          'text-(--popup-window-delimiter-text-color)',
+          'font-(--popup-window-delimiter-font-weight)',
+          'after:ms-[10px] after:block after:flex-1 after:min-w-[15px] after:h-px after:bg-(--popup-window-delimiter-bg-color)'
         ].join(' '),
-        // @todo fix this
-        separator: '-mx-1 my-1 h-px bg-base-master/10',
-        // @todo fix this
+        separator: 'my-[8px] mx-[18px] h-[1px] bg-(--popup-window-delimiter-bg-color)',
         item: [
           'group',
-          'w-full',
+          'w-full min-w-[195px] h-[36px]',
+          'px-[18px] mt-(--menu-item-block-stack-space)',
           'relative',
-          'flex items-center',
-          'select-none outline-none',
+          'flex flex-row rtl:flex-row-reverse items-center',
+          'select-none outline-none whitespace-nowrap',
           'cursor-pointer',
-          'before:absolute before:z-[-1] before:inset-px before:rounded-2xs',
           'data-disabled:cursor-not-allowed data-disabled:opacity-75',
-          'text-base-master',
-          'data-highlighted:not-data-disabled:text-base-900 data-highlighted:not-data-disabled:before:bg-base-100/50',
-          'data-[state=checked]:text-base-900 data-[state=checked]:before:bg-base-100/50',
-          'transition-colors',
-          // from dropdown-menu
+          'text-start',
+          'text-(length:--menu-popup-item-font-size)',
           'text-(--menu-popup-item-color) hover:text-(--menu-popup-item-color-hover)',
-          'data-highlighted:text-(--menu-popup-item-color-hover)',
-          // 'data-[state=open]:text-(--menu-popup-item-color-hover)',
+          'data-highlighted:not-data-disabled:text-(--menu-popup-item-color-hover)',
+          'data-[state=open]:text-(--menu-popup-item-color-hover)',
           'hover:bg-(--menu-popup-item-bg-color-hover)',
           'data-highlighted:bg-(--menu-popup-item-bg-color-hover)',
-          // 'data-[state=open]:bg-(--menu-popup-item-bg-color-hover)',
+          'data-[state=open]:bg-(--menu-popup-item-bg-color-hover)',
+          'transition-colors',
+          // @todo fix this
           // from list colors for dropdown
           ''
         ].join(' '),
-        // @todo fix this
         itemLeadingIcon: [
           'shrink-0',
           'size-[18px]',
           'text-(--ui-color-design-plain-content-icon-secondary)',
           'group-data-highlighted:not-data-disabled:text-(--ui-color-accent-main-primary)',
-          // 'group-data-[state=open]:text-(--ui-color-accent-main-primary)',
+          'group-data-[state=open]:text-(--ui-color-accent-main-primary)',
           'group-data-[state=checked]:text-(--ui-color-accent-main-primary)',
           'transition-colors'
         ].join(' '),
-        itemLeadingAvatar: 'shrink-0',
-        itemLeadingAvatarSize: '',
-        itemLeadingChip: 'shrink-0',
-        itemLeadingChipSize: '',
-        itemTrailing: 'ms-auto inline-flex gap-1.5 items-center',
-        itemTrailingIcon: 'shrink-0',
-        itemLabel: 'truncate'
+        itemLeadingAvatar: 'shrink-0 size-[16px] mx-px', // @memo 18-2px
+        itemLeadingAvatarSize: '2xs', // @memo this wrong
+        itemLeadingChip: 'shrink-0 size-[16px] mx-px',
+        itemLeadingChipSize: 'sm',
+        itemTrailing: 'ml-auto rtl:ml-0 rtl:mr-auto inline-flex gap-1.5 items-center',
+        itemTrailingIcon: 'shrink-0 size-[24px] text-(--ui-color-design-plain-content-icon-secondary)',
+        itemLabel: [
+          'truncate ms-[2px] -mt-px',
+          'group-data-[state=checked]:text-(--ui-color-accent-main-primary)'
+        ].join(' ')
       },
       variants: {
         ...buttonGroupVariantWithRoot,
         size: {
-          xss: defSize,
-          xs: defSize,
-          sm: defSize,
-          md: defSize,
-          lg: defSize,
-          xl: defSize
+          xss: {
+            base: 'h-[20px] gap-1 text-(length:--ui-font-size-4xs)/[normal]',
+            leading: 'px-1',
+            trailing: 'px-1',
+            leadingIcon: 'size-[12px]',
+            leadingAvatarSize: '3xs',
+            trailingIcon: 'size-[12px]'
+          },
+          xs: {
+            base: 'h-[24px] gap-1 text-(length:--ui-font-size-xs)/[normal]',
+            leading: 'px-1',
+            trailing: 'px-1',
+            leadingIcon: 'size-[14px]',
+            leadingAvatarSize: '3xs',
+            trailingIcon: 'size-[14px]'
+          },
+          sm: {
+            base: 'h-[28px] gap-1.5 text-(length:--ui-font-size-sm)/[normal]',
+            leading: 'px-1.5',
+            trailing: 'px-1.5',
+            leadingIcon: 'size-[16px]',
+            leadingAvatar: 'size-[16px]',
+            leadingAvatarSize: '2xs',
+            trailingIcon: 'size-[16px]'
+          },
+          md: {
+            base: 'h-[34px] gap-1.5 text-(length:--ui-font-size-lg)/[normal]',
+            leading: 'px-2',
+            trailing: 'px-2',
+            leadingIcon: 'size-[18px]',
+            leadingAvatarSize: '2xs',
+            trailingIcon: 'size-[18px]'
+          },
+          lg: {
+            base: 'h-[38px] gap-2 text-(length:--ui-font-size-lg)/[normal]',
+            leading: 'px-2',
+            trailing: 'px-2',
+            leadingIcon: 'size-[22px]',
+            leadingAvatarSize: '2xs',
+            trailingIcon: 'size-[22px]'
+          },
+          xl: {
+            base: 'h-[46px] gap-2 text-(length:--ui-font-size-2xl)/[normal]',
+            leading: 'px-2',
+            trailing: 'px-2',
+            leadingIcon: 'size-[22px]',
+            leadingAvatarSize: '2xs',
+            trailingIcon: 'size-[22px]'
+          }
+        },
+        colorItem: {
+          'air-primary': { item: 'style-filled' },
+          'air-primary-success': { item: 'style-filled-success' },
+          'air-primary-alert': { item: 'style-filled-alert' },
+          'air-primary-copilot': { item: 'style-filled-copilot' },
+          'air-primary-warning': { item: 'style-filled-warning' },
+          'default': { item: 'style-old-default' },
+          'danger': { item: 'style-old-danger' },
+          'success': { item: 'style-old-success' },
+          'warning': { item: 'style-old-warning' },
+          'primary': { item: 'style-old-primary' },
+          'secondary': { item: 'style-old-secondary' },
+          'collab': { item: 'style-old-collab' },
+          'ai': { item: 'style-old-ai' }
         }
-        // ...colorDropDownItem
       },
       compoundVariants: [
         // from dropdown-menu
         {
-          color: ['air-primary', 'air-primary-success', 'air-primary-alert', 'air-primary-copilot', 'air-primary-warning'],
+          colorItem: ['air-primary', 'air-primary-success', 'air-primary-alert', 'air-primary-copilot', 'air-primary-warning'],
           active: false,
           class: {
             item: [
@@ -158,7 +203,7 @@ export default () => {
           }
         },
         {
-          color: ['air-primary', 'air-primary-success', 'air-primary-alert', 'air-primary-copilot', 'air-primary-warning'],
+          colorItem: ['air-primary', 'air-primary-success', 'air-primary-alert', 'air-primary-copilot', 'air-primary-warning'],
           active: true,
           class: {
             item: [
@@ -170,17 +215,17 @@ export default () => {
           }
         },
         {
-          color: 'default',
+          colorItem: 'default',
           active: false,
           class: ''
         },
         {
-          color: 'default',
+          colorItem: 'default',
           active: true,
           class: ''
         },
         {
-          color: ['danger', 'success', 'warning', 'primary', 'secondary', 'collab', 'ai'],
+          colorItem: ['danger', 'success', 'warning', 'primary', 'secondary', 'collab', 'ai'],
           active: false,
           class: {
             item: [
@@ -192,7 +237,7 @@ export default () => {
           }
         },
         {
-          color: ['danger', 'success', 'warning', 'primary', 'secondary', 'collab', 'ai'],
+          colorItem: ['danger', 'success', 'warning', 'primary', 'secondary', 'collab', 'ai'],
           active: true,
           class: {
             item: [