+++ src/theme/tools/color-drop-down-item.ts
@@ -5,45 +5,18 @@
  */
 export default {
   colorItem: {
-    /**
-     * @memo not fill - use default data
-     */
-    default: '',
-    danger: {
-      // item: [
-      //   'text-red-900',
-      //   'data-highlighted:text-red-900 data-highlighted:before:bg-red-100',
-      //   'data-[state=checked]:text-red-900 data-[state=checked]:before:bg-red-100'
-      // ].join(' '),
-      // itemLeadingIcon: [
-      //   'text-red-600',
-      //   'group-data-highlighted:text-red-900',
-      //   'group-data-[state=checked]:text-red-900'
-      // ].join(' ')
-    },
-    success: {
-      item: '',
-      itemLeadingIcon: ''
-    },
-    warning: {
-      item: '',
-      itemLeadingIcon: ''
-    },
-    primary: {
-      item: '',
-      itemLeadingIcon: ''
-    },
-    secondary: {
-      item: '',
-      itemLeadingIcon: ''
-    },
-    collab: {
-      item: '',
-      itemLeadingIcon: ''
-    },
-    ai: {
-      item: '',
-      itemLeadingIcon: ''
-    }
+    'air-primary': { item: 'style-filled' },
+    'air-primary-success': { item: 'style-filled-success' },
+    'air-primary-alert': { item: 'style-filled-alert' },
+    'air-primary-copilot': { item: 'style-filled-copilot' },
+    'air-primary-warning': { item: 'style-filled-warning' },
+    'default': { item: 'style-old-default' },
+    'danger': { item: 'style-old-danger' },
+    'success': { item: 'style-old-success' },
+    'warning': { item: 'style-old-warning' },
+    'primary': { item: 'style-old-primary' },
+    'secondary': { item: 'style-old-secondary' },
+    'collab': { item: 'style-old-collab' },
+    'ai': { item: 'style-old-ai' }
   }
 }